+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Better highlight the position picker input-box !1583 !1586 @ChrisOelmueller
 - Removed faq and replaced to external freshdesk support form in top and footer menu #817 !1587 @chriswalg
 - Call foodsaver from pickup slot dropdown menu #772 !1591 @ChrisOelmueller
+- Hide all phone numbers for unverified store team members and jumpers !1591 @ChrisOelmueller
 
 ## Bugfixes
 - Clarify that new forum threads won't reach members inactive for more than six months !1553 ("merging" !1385 and !1233) @zommuter